+++ packages/callhierarchy/src/browser/callhierarchy-frontend-module.ts
@@ -20,14 +20,11 @@ import { CallHierarchyService, CallHierarchyServiceProvider } from './callhierar
 import { WidgetFactory, bindViewContribution } from '@theia/core/lib/browser';
 import { CALLHIERARCHY_ID } from './callhierarchy';
 import { createHierarchyTreeWidget } from './callhierarchy-tree';
-import { CurrentEditorAccess } from './current-editor-access';
 import { ContainerModule } from '@theia/core/shared/inversify';
 
 import '../../src/browser/style/index.css';
 
 export default new ContainerModule(bind => {
-    bind(CurrentEditorAccess).toSelf().inSingletonScope();
-
     bindContributionProvider(bind, CallHierarchyService);
     bind(CallHierarchyServiceProvider).to(CallHierarchyServiceProvider).inSingletonScope();