+++ packages/core/src/browser/view-container.ts
@@ -912,10 +912,7 @@ export class ViewContainerPart extends BaseWidget {
     protected readonly body: HTMLElement;
     protected readonly collapsedEmitter = new Emitter<boolean>();
     protected readonly contextMenuEmitter = new Emitter<MouseEvent>();
-    /**
-     * @deprecated since 0.11.0, use `onDidChangeVisibility` instead
-     */
-    readonly onVisibilityChanged = this.onDidChangeVisibility;
+
     protected readonly onTitleChangedEmitter = new Emitter<void>();
     readonly onTitleChanged = this.onTitleChangedEmitter.event;
     protected readonly onDidFocusEmitter = new Emitter<this>();