+++ packages/terminal/src/browser/terminal-theme-service.ts
@@ -16,7 +16,8 @@
 
 import { ITheme } from 'xterm';
 import { injectable, inject } from '@theia/core/shared/inversify';
-import { ColorRegistry, ColorDefaults } from '@theia/core/lib/browser/color-registry';
+import { ColorRegistry } from '@theia/core/lib/browser/color-registry';
+import { ColorDefaults } from '@theia/core/lib/common/color';
 import { ThemeService } from '@theia/core/lib/browser/theming';
 
 /**